+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Drag & drop file upload #653
+- Disallow use of "any" type #624
 - Date picker #559
 - New login page #573
 - Autocomplete name field in DOI form #529
@@ -67,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Cast integer fields as numbers #661
 - Disallow object update with invalid form #656
 - Handle required form array fields #484
 - Prevent user deleting object in published folder #486